requirements.txt
# Core data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Korean market data
OpenDartReader>=0.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.31.0
websockets>=11.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine learning and AI
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0
langchain>=0.0.200

# Natural language processing
nltk>=3.8.0
textblob>=0.17.0

# Technical analysis
ta-lib>=0.4.0
yfinance>=0.2.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Caching and performance
redis>=4.6.0
celery>=5.3.0

# Security
cryptography>=41.0.0
passlib>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0 